tests/dev-requirements.txt
pytest-playwright>=0.7.0
